Speed up the PIC32MX795F512L to the full 80MHz.
PLL settings change, as well as the PBCLK divider. FPBCLK now 40000000 on the faster chip. CPU is about 4.5 times faster than the other 48MHz parts, due to the cache and prefetch buffer as well as the higher clock speed. LED and SQWAVE timing is unaffected because that's all derived from PBCLK.